Extract features list to a dedicated component: FeaturesList
