+++ containers/payments/subscription/FeaturesList.js
@@ -1,45 +1,21 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { c } from 'ttag';
-import { useUser, useOrganization, Href, Icon } from 'react-components';
-import { Link } from 'react-router-dom';
+import { Icon } from 'react-components';
 
-const FeaturesList = () => {
-    const [{ hasPaidMail } = {}] = useUser();
-    const [{ MaxMembers } = {}] = useOrganization();
-    const features = [
-        ...(hasPaidMail
-            ? [
-                  { to: '/settings/auto-reply', text: c('Link').t`Add auto-reply` },
-                  { to: '/settings/domains', text: c('Link').t`Add custom domain` },
-                  { to: '/settings/filters', text: c('Link').t`Add filters` },
-                  { to: '/contacts', text: c('Link').t`Manage encrypted contacts`, internal: true }
-              ]
-            : []),
-        ...(MaxMembers > 1 ? [{ to: '/settings/members', text: c('Link').t`Add new users` }] : []),
-        { to: 'https://protonvpn.com/download/', text: c('Link').t`Use ProtonVPN`, external: true }
-    ];
+const FeaturesList = ({ features = [] }) => {
+    if (!features.length) {
+        return null;
+    }
     return (
         <ul className="unstyled flex flex-nowrap flex-spacebetween">
-            {features.map(({ to, text, external, internal }, index) => {
+            {features.map((text, index) => {
                 const key = `${index}`;
                 return (
-                    <li key={key} className="mb0-5 aligncenter">
-                        {external || internal ? (
-                            <Href href={to} target={external ? '_blank' : '_self'}>
-                                <div>
-                                    <Icon name="add" size={25} />
-                                </div>
-                                {text}
-                            </Href>
-                        ) : (
-                            <Link to={to}>
-                                <div>
-                                    <Icon name="add" size={25} />
-                                </div>
-                                {text}
-                            </Link>
-                        )}
+                    <li key={key} className="aligncenter pl0-5 pr0-5">
+                        <div>
+                            <Icon name="add" size={25} className="fill-pm-blue" />
+                        </div>
+                        {text}
                     </li>
                 );
             })}
@@ -48,7 +24,7 @@ const FeaturesList = () => {
 };
 
 FeaturesList.propTypes = {
-    model: PropTypes.object
+    features: PropTypes.arrayOf(PropTypes.string)
 };
 
 export default FeaturesList;
